Move UpdateLastSeen logic into helper
