+++ src/lib/ui/uiMutationHelper.ts
@@ -1,5 +1,8 @@
+import {browser} from '$app/env';
 import type {Entity} from '$lib/vocab/entity/entity';
-import {derived, writable, type Writable} from '@feltcoop/svelte-gettable-stores';
+import type {DirectoryEntityData} from '$lib/vocab/entity/entityData';
+import {derived, writable, type Readable, type Writable} from '@feltcoop/svelte-gettable-stores';
+import {LAST_SEEN_KEY} from './app';
 import type {WritableUi} from './ui';
 
 export const setFreshnessDerived = (ui: WritableUi, directory: Writable<Entity>): void => {
@@ -25,5 +28,36 @@ export const upsertCommunityFreshnessById = (ui: WritableUi, community_id: numbe
 		freshnessByCommunityId.get(community_id)!.set(fresh);
 	} else {
 		freshnessByCommunityId.set(community_id, writable(fresh));
+	}
+};
+
+export const updateLastSeen = (
+	ui: WritableUi,
+	directory_id: number,
+	time: number | undefined,
+): void => {
+	const {lastSeenByDirectoryId, entityById, spaceById} = ui;
+	const timestamp = time ?? Date.now();
+
+	if (lastSeenByDirectoryId.has(directory_id)) {
+		lastSeenByDirectoryId.get(directory_id)!.set(timestamp);
+	} else {
+		lastSeenByDirectoryId.set(directory_id, writable(timestamp));
+	}
+
+	if (browser) {
+		localStorage.setItem(`${LAST_SEEN_KEY}${directory_id}`, `${timestamp}`);
+	}
+
+	const directory = entityById.get(directory_id) as
+		| Readable<Entity & {data: DirectoryEntityData}>
+		| undefined;
+
+	if (directory) {
+		upsertCommunityFreshnessById(
+			ui,
+			//TODO add directory field to space & vice versa to avoid this mess below
+			spaceById.get(directory.get().data.space_id)!.get().community_id,
+		);
 	}
 };
